Directions: Auto recalculate direction on marker drag
